% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hazard_rate.R
\name{adjust_survival_curve}
\alias{adjust_survival_curve}
\title{Adjust Hazards Underlying a Survival Curve}
\usage{
adjust_survival_curve(
  baseline_survival,
  treatment_survival,
  adjustment_mode = "baseline_after_time",
  adjustment_time = 1
)
}
\arguments{
\item{baseline_survival}{A numeric vector of survival curve for the baseline
group.}

\item{treatment_survival}{A numeric vector of survival curve for the
treatment group. `treatment_survival` is expected to have an equal length to
`baseline_survival`.}

\item{adjustment_mode}{A character string specifying the adjustment mode.
Options are:
  \itemize{
    \item `"baseline_after_time"` (default): Adjust hazard rate to match
    baseline at and beyond the specified time point.
    \item `"higher_HR_after_time"`: Use the higher hazard rate between
    treatment and baseline after the specified time point.
    \item `"lower_HR_after_time"`: Use the lower hazard rate between
    treatment and baseline after the specified time point.
  }}

\item{adjustment_time}{A numeric value representing the time point at which
to start adjusting the hazard rate for the treatment group. Default is `1`.
`adjustment_time` expects each element in the `baseline_survival` and
`treatment_survival` survival curves (vectors) to represent the proportion of
survivors at a time point. The scale or units of the values passed to the
`adjustment_time` argument are implicitly equivalent to those corresponding
to the survival curves.}
}
\value{
A numeric vector of adjusted hazard rates for the treatment group.
}
\description{
This function compares two survival curves representing a baseline group and
a treatment group, and adjusts the hazard rate for the treatment group at and
beyond a specified time point to match the hazard rate for the baseline group
or to follow specified adjustment mode.
}
\examples{
\dontrun{
# Generate simulated survival data for a baseline and a treatment group
set.seed(123)
days_in_years <- 1:(365*3)
baseline_survival <- 1 - pweibull(q = days_in_years, shape = 1.1, scale = 650)
treatment_survival <- 1 - pweibull(q = days_in_years, shape = 4, scale = 900)

# Compare the simulated survival data
plot(
  x = days_in_years/365,
  y = baseline_survival,
  ylim = c(0, 1),
  type = "l",
  col = "red",
  xlab = "Time in years",
  ylab = "Survival"
 )
lines(x = days_in_years/365, treatment_survival, col = "blue")

# Estimate and compare the hazard rates for both groups
baseline_hazard <- convert_survival_to_hazard(baseline_survival)
treatment_hazard <- convert_survival_to_hazard(treatment_survival)

plot(
  x = days_in_years/365,
  y = baseline_hazard,
  ylim = c(min(treatment_hazard), c(max(treatment_hazard))),
  type = "l",
  col = "red",
  xlab = "Time in years",
  ylab = "Hazard rate"
 )
lines(x = days_in_years/365, treatment_hazard, col = "blue")

# Adjust the hazard rate for the treatment group to match the hazard rate for
# the baseline group at and beyond time point 365
adjusted_survival <- NeuroblastomaPSM::adjust_survival_curve(
  baseline_survival = baseline_survival,
  treatment_survival = treatment_survival,
  adjustment_time = 365
)

plot(
  x = days_in_years/365,
  y = treatment_survival,
  type = "l",
  col = "red",
  xlab = "Time",
  ylab = "Survival"
)
lines(x = days_in_years/365, y = baseline_survival, col = "blue")
lines(x = days_in_years/365, y = adjusted_survival, col = "green")

# Adjust the hazard rate for the treatment group to the higher value compared
# to the baseline group after time point 365
adjusted_survival <- NeuroblastomaPSM::adjust_survival_curve(
  baseline_survival = baseline_survival,
  treatment_survival = treatment_survival,
  adjustment_mode = "higher_HR_after_time",
  adjustment_time = 365
)

plot(
  x = days_in_years/365,
  y = treatment_survival,
  type = "l",
  col = "red",
  xlab = "Time in years",
  ylab = "Survival"
)
lines(x = days_in_years/365, y = baseline_survival, col = "blue")
lines(x = days_in_years/365, y = adjusted_survival, col = "green")

# Adjust the hazard rate for the treatment group to the lower value compared
# to the baseline group after time point 365
adjusted_survival <- NeuroblastomaPSM::adjust_survival_curve(
  baseline_survival = baseline_survival,
  treatment_survival = treatment_survival,
  adjustment_mode = "lower_HR_after_time",
  adjustment_time = 365
)

plot(
  x = days_in_years/365,
  y = treatment_survival,
  type = "l",
  col = "red",
  xlab = "Time in years",
  ylab = "Survival"
)
lines(x = days_in_years/365, y = baseline_survival, col = "blue")
lines(x = days_in_years/365, y = adjusted_survival, col = "green")
}
}
