% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_economic_analysis.R
\name{calculate_ae_costs}
\alias{calculate_ae_costs}
\title{Calculate Adverse Effects (AE) Costs}
\usage{
calculate_ae_costs(params)
}
\arguments{
\item{params}{A list of model parameters including the list of parameters in
the list \code{\link{l_psm_parameters}} in addition to:
\itemize{
  \item \code{time_horizon}: The time horizon for the model in years.
  \item \code{cycle_length}: The length of a model cycle measured in years.
  \item \code{disc_rate_costs}: The annual discount rate for incurred costs.
  \item \code{disc_rate_qalys}: The annual discount rate for accrued QALYs.
}}
}
\value{
A scalar representing the costs of related to `Dinutuximab β` (GD2)
adverse events.
}
\description{
This function calculates the costs associated with the adverse effects of
receiving `Dinutuximab β` (GD2) over a given time horizon, based on the
provided model parameters.
}
\examples{
\dontrun{
# Define model parameters
params <- c(
  time_horizon = 10,
  cycle_length = 1/12,
  disc_rate_costs = 0.035,
  disc_rate_qalys = 0.015,
  NeuroblastomaPSM::l_psm_parameters
)

# Calculate AE costs
GD2_ae_costs <- calculate_ae_costs(params = params)

GD2_ae_costs
}
}
