% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_economic_analysis.R
\name{calculate_efs_costs}
\alias{calculate_efs_costs}
\title{Calculate Event Free Survival (EFS) Costs}
\usage{
calculate_efs_costs(params, GD2_cycle_days, TT_cycle_days)
}
\arguments{
\item{params}{A list of model parameters including the list of parameters in
the list \code{\link{l_psm_parameters}} in addition to:
\itemize{
  \item \code{time_horizon}: The time horizon for the model in years.
  \item \code{cycle_length}: The length of a model cycle measured in years.
  \item \code{disc_rate_costs}: The annual discount rate for incurred costs.
  \item \code{disc_rate_qalys}: The annual discount rate for accrued QALYs.
}}

\item{GD2_cycle_days}{The number of days in a GD2 treatment cycle. Default
value is `35` days.}

\item{TT_cycle_days}{The number of days in a TT treatment cycle. Default
value is `30` days.}
}
\value{
A list containing `Dinutuximab β` (GD2) and `Isotretinoin` (TT) costs
over the model time-horizon.
}
\description{
This function calculates the costs associated with the Neuroblastoma
interventions, `Dinutuximab β` (GD2) and `Isotretinoin` (TT).
}
\examples{
\dontrun{
# Define model parameters
params <- c(
  time_horizon = 10,
  cycle_length = 1/12,
  disc_rate_costs = 0.035,
  disc_rate_qalys = 0.015,
  NeuroblastomaPSM::l_psm_parameters
)

# Calculate AE treatment costs
l_GD2_efs_costs <- calculate_efs_costs(
 params = params,
 GD2_cycle_days = 35,
 TT_cycle_days = 30
)

l_GD2_efs_costs
}
}
