% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_markov_trace.R
\name{calculate_markov_trace}
\alias{calculate_markov_trace}
\title{Calculate Markov Trace}
\usage{
calculate_markov_trace(df_survival_curves_long)
}
\arguments{
\item{df_survival_curves_long}{A data frame containing the predicted
cumulative survival curves in long format with columns for time, treatment,
end_point, and survival probabilities.}
}
\value{
A data frame in wide format with columns for time, treatment, and
states occupancy (`EFS`, `PPS`, `D`).
}
\description{
This function calculates the state membership (Markov trace) for different
health states based on predicted survival curves.
}
\examples{
\dontrun{
# Load the fitted Gompertz model parameters
models_fit <- NeuroblastomaPSM::parametric_models

# Define model parameters
params <- list(
  time_horizon = 10,
  cycle_length = 1/12,
  body_weight = 15,
  GD2_unit_mg = 20,
  GD2_unit_price = 2000,
  GD2_dose_days = 10,
  TT_unit_mg = 10,
  TT_unit_price = 3000,
  TT_dose_days = 14,
  GD2_aEvents_names = c("fever", "diarrhea", "vomiting", "infection",
    "hypersensitivity_reaction", "capillary_leak"),
  GD2_aEvents_probs = c(
    "fever" = 0.18,
    "diarrhea" = 0.07,
    "vomiting" = 0.06,
    "infection" = 0.21,
    "hypersensitivity_reaction" = 0.11,
    "capillary_leak" = 0.06
  ),
  GD2_aEvents_costs = c(
    "fever" = 1.98,
    "diarrhea" = 3.2,
    "vomiting" = 8.45,
    "infection" = 222.33,
    "hypersensitivity_reaction" = 4.13,
    "capillary_leak" = 307.37
  ),
  Temo_unit_mg = 50,
  Temo_unit_price = 2000,
  Temo_dose_days = 5,
  Iri_unit_mg = 40,
  Iri_unit_price = 3000,
  Iri_dose_days = 5,
  u_EFS = 0.23,
  u_PPS = 0.23,
  disc_rate_costs = 0.035,
  disc_rate_qalys = 0.015
)

# Predict cumulative survival
df_survival_curves_long <- NeuroblastomaPSM::predict_cumulative_survival(
  models_fit = models_fit,
  params = params
)

# Generate Markov trace
df_markov_trace <- NeuroblastomaPSM::calculate_markov_trace(
  df_survival_curves_long = df_survival_curves_long
)

rbind(
  head(df_markov_trace, n = 5),
  tail(df_markov_trace, n = 5)
)
}
}
