% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_economic_analysis.R
\name{calculate_pps_costs}
\alias{calculate_pps_costs}
\title{Calculate Post Progression Survival (PPS) Costs}
\usage{
calculate_pps_costs(params, Temo_cycle_days, Iri_cycle_days)
}
\arguments{
\item{params}{A list of model parameters including the list of parameters in
the list \code{\link{l_psm_parameters}} in addition to:
\itemize{
  \item \code{time_horizon}: The time horizon for the model in years.
  \item \code{cycle_length}: The length of a model cycle measured in years.
  \item \code{disc_rate_costs}: The annual discount rate for incurred costs.
  \item \code{disc_rate_qalys}: The annual discount rate for accrued QALYs.
}}

\item{Temo_cycle_days}{The number of days in a Temozolomide treatment cycle.
Default value is `21` days.}

\item{Iri_cycle_days}{The number of days in a Irinotecan treatment cycle.
Default value is `21` days.}
}
\value{
A scalar representing the costs associated with PPS.
}
\description{
Calculate Post Progression Survival (PPS) Costs
}
\examples{
\dontrun{
# Define model parameters
params <- c(
  time_horizon = 10,
  cycle_length = 1/12,
  disc_rate_costs = 0.035,
  disc_rate_qalys = 0.015,
  NeuroblastomaPSM::l_psm_parameters
)

# Calculate treatment costs
pps_costs <- calculate_pps_costs(
 params = params,
 Temo_cycle_days = 21,
 Iri_cycle_days = 21
)

pps_costs
}
}
