% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_economic_analysis.R
\name{calculate_treatment_costs}
\alias{calculate_treatment_costs}
\title{Calculate Treatment Costs}
\usage{
calculate_treatment_costs(
  params,
  GD2_cycle_days = 35,
  TT_cycle_days = 30,
  Temo_cycle_days = 21,
  Iri_cycle_days = 21
)
}
\arguments{
\item{params}{A list of model parameters including the list of parameters in
the list \code{\link{l_psm_parameters}} in addition to:
\itemize{
  \item \code{time_horizon}: The time horizon for the model in years.
  \item \code{cycle_length}: The length of a model cycle measured in years.
  \item \code{disc_rate_costs}: The annual discount rate for incurred costs.
  \item \code{disc_rate_qalys}: The annual discount rate for accrued QALYs.
}}

\item{GD2_cycle_days}{The number of days in a GD2 treatment cycle. Default
value is `35` days.}

\item{TT_cycle_days}{The number of days in a TT treatment cycle. Default
value is `30` days.}

\item{Temo_cycle_days}{The number of days in a Temozolomide treatment cycle.
Default value is `21` days.}

\item{Iri_cycle_days}{The number of days in a Irinotecan treatment cycle.
Default value is `21` days.}
}
\value{
A matrix of treatment costs for the first year, with columns for GD2
and TT costs over the specified time points.
}
\description{
This function calculates the treatment costs, , `Dinutuximab β` (GD2) and
`Isotretinoin` (TT). including the costs of managing any associated adverse
events, over a given time horizon.
}
\examples{
\dontrun{
# Define model parameters
params <- c(
  time_horizon = 10,
  cycle_length = 1/12,
  disc_rate_costs = 0.035,
  disc_rate_qalys = 0.015,
  NeuroblastomaPSM::l_psm_parameters
)

# Calculate treatment costs
l_treatment_costs <- calculate_treatment_costs(params = params)

View(l_treatment_costs)
}
}
