% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hazard_rate.R
\name{check_all_equal_tol}
\alias{check_all_equal_tol}
\title{Check if all values of a vector are equal to the first value with a given
tolerance}
\usage{
check_all_equal_tol(x, tolerance)
}
\arguments{
\item{x}{A numeric vector to check.}

\item{tolerance}{The tolerance to use when checking for equality.}
}
\value{
TRUE if all values of x are equal to the first value with the given
tolerance, FALSE otherwise.
}
\description{
This function checks if all values of a numeric vector are equal to the first
value with a given tolerance.
}
\examples{
\dontrun{

check_all_equal_tol(c(3, 3, 3, 3, 3), 0.00001)
# returns TRUE

check_all_equal_tol(c(3, 3, 3, 3, 3.0001), 0.00001)
# returns FALSE

}
}
