% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hazard_rate.R
\name{convert_survival_to_hazard}
\alias{convert_survival_to_hazard}
\title{Convert Survival Curve to Cumulative Hazard Rates}
\usage{
convert_survival_to_hazard(survival_curve)
}
\arguments{
\item{survival_curve}{A numeric vector representing survival probabilities.}
}
\value{
A numeric vector of instantaneous hazard rates.
}
\description{
This function converts a numeric vector of survival probabilities into
instantaneous hazard rates. The relationship between survival probabilities
and the discrete hazard rates is given by:
}
\details{
\deqn{h(t_i) = -\log\left(\frac{S(t_i)}{S(t_{i-1})}\right)}

where \eqn{S(t_i)} is the survival probability at time \eqn{t_i} and
\eqn{h(t_i)} is the hazard rate at time \eqn{t_i}.
}
\examples{
\dontrun{
survival_curve <- c(0.99, 0.95, 0.90, 0.85)

cum_hazard_rates <- NeuroblastomaPSM::convert_survival_to_hazard(
survival_curve = survival_curve
)

cum_hazard_rates
}
}
