% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_psa.R
\name{estimate_se}
\alias{estimate_se}
\title{Estimate Standard Error from Mean and 95\% Confidence Interval}
\usage{
estimate_se(mean, lower_ci, upper_ci, distribution)
}
\arguments{
\item{mean}{A numeric value representing the mean of the distribution.}

\item{lower_ci}{A numeric value representing the lower bound of the 95\%
confidence interval.}

\item{upper_ci}{A numeric value representing the upper bound of the 95\%
confidence interval.}

\item{distribution}{A character string specifying the type of distribution.
Options are "normal", "beta", and "gamma".}
}
\value{
A named list containing the mean and a value representing the
estimated standard error (SE), upper and lower bounds of the 95\% confidence
interval, and the parameter distribution.
}
\description{
This function estimates the standard error (SE) from the mean and 95\%
confidence interval (CI) for normal, beta, and gamma distributions.
}
\details{
The standard error (SE) is calculated using different formulas depending on
the specified distribution:

\describe{
  \item{Normal Distribution}{
    The margin of error is calculated as \eqn{(upper\_ci - lower\_ci) / 2}.
    The standard error is then calculated as:
    \deqn{SE = \frac{margin\_of\_error}{1.96}}
  }
  \item{Beta Distribution}{
    First, the parameters \eqn{\alpha} and \eqn{\beta} are estimated using
    the method of moments:
    \deqn{\alpha = mean \times \left( \frac{mean \times (1 - mean)}
    {\left( \frac{margin\_of\_error}{1.96} \right)^2} - 1 \right)}
    \deqn{\beta = \alpha \times \left( \frac{1}{mean} - 1 \right)}
    The standard error is then calculated as:
    \deqn{SE = \sqrt{\frac{\alpha \times \beta}{(\alpha + \beta)^2 \times
    (\alpha + \beta + 1)}}}
  }
  \item{Gamma Distribution}{
    First, the shape parameter \eqn{k} and the scale parameter \eqn{\theta}
    are estimated:
    \deqn{k = \frac{mean^2}{(margin\_of\_error / 1.96)^2}}
    \deqn{\theta = \frac{(margin\_of\_error / 1.96)^2}{mean}}
    The standard error is then calculated as:
    \deqn{SE = \frac{\sqrt{k} \times \theta}{\sqrt{mean}}}
  }
}
}
\examples{
\dontrun{
# Example usage for normal distribution
mean <- 50
lower_ci <- 45
upper_ci <- 55
distribution <- "normal"
se <- estimate_se(mean, lower_ci, upper_ci, distribution)
print(paste("Estimated Standard Error for", distribution, "distribution:", se))

# Example usage for beta distribution
mean <- 0.5
lower_ci <- 0.4
upper_ci <- 0.6
distribution <- "beta"
se <- estimate_se(mean, lower_ci, upper_ci, distribution)
print(paste("Estimated Standard Error for", distribution, "distribution:", se))

# Example usage for gamma distribution
mean <- 10
lower_ci <- 8
upper_ci <- 12
distribution <- "gamma"
se <- estimate_se(mean, lower_ci, upper_ci, distribution)
print(paste("Estimated Standard Error for", distribution, "distribution:", se))
}
}
