% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_survival_curve.R
\name{get_lifeTable_mortality_probs}
\alias{get_lifeTable_mortality_probs}
\title{Estimate Mortality Probabilities from a Survival Curve}
\usage{
get_lifeTable_mortality_probs(survival_curve)
}
\arguments{
\item{survival_curve}{A numeric vector representing the survival curve
corresponding to each age.}
}
\value{
A numeric vector of estimated mortality probabilities corresponding
to each age.
}
\description{
This function estimates the mortality probabilities from a given cumulative
survival curve. The survival curve represents the probability of beyond each
time point or age.
}
\examples{
\dontrun{
mortality_probs <- c(0.01, 0.02, 0.03, 0.04)

survival_curve <- get_lifeTable_survival_curve(mortality_probs)

estimated_mortality_probs <- get_lifeTable_mortality_probs(survival_curve)

estimated_mortality_probs
}
}
