% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hazard_rate.R
\name{get_lifeTable_survival_curve}
\alias{get_lifeTable_survival_curve}
\title{Calculate Survival Curve from Mortality Probabilities}
\usage{
get_lifeTable_survival_curve(mortality_probs)
}
\arguments{
\item{mortality_probs}{A numeric vector representing the mortality
probabilities for each age or time point.}
}
\value{
A numeric vector representing the survival probabilities at each age
or time point.
}
\description{
This function calculates the survival curve from a given vector of mortality
probabilities. The survival curve represents the probability of surviving up
to each age or time point, starting from a cohort with an initial size of 1
(or 100% survival).
}
\details{
The survival curve (\eqn{S(t)}) is calculated as the cumulative product of
(1 - mortality probability) for each time period.

\deqn{S(t) = \prod_{i=1}^{t} (1 - q_i)}

where \eqn{q_i} is the mortality probability at time \eqn{i}.
}
\examples{
\dontrun{
# Define a vector of mortality probabilities
mortality_probs <- df_lifeTable$`Both sexes`

# Calculate the survival curve
survival_curve <- NeuroblastomaPSM::get_lifeTable_survival_curve(
   mortality_probs = mortality_probs
)

survival_curve
plot(survival_curve, type = 'l')
}
}
