% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPD_data}
\alias{IPD_data}
\title{Reconstructed Individual Patient Data}
\format{
A data frame with 624 observations on the following 5 variables:
\describe{
  \item{trt}{Character vector indicating the treatment received by each
  patient. Possible values are \code{"Isotretinoin"} and
  \code{"Dinutuximab β"}.}
  \item{trt_cd}{Character vector representing the treatment code.
  \code{"TT"} for \code{Isotretinoin} and \code{"GD2"} for
  \code{Dinutuximab β}.}
  \item{curve}{Character vector indicating the type of survival curve.
  Possible values are \code{"OS"} (Overall Survival) and \code{"EFS"}
  (Event-Free Survival).}
  \item{eventtime}{Numeric vector representing the time to event or censoring
  for each patient.}
  \item{event}{Integer vector where \code{1} indicates the occurrence of the
  event and \code{0} indicates censoring.}
}
}
\source{
{Khader to add the reference!}
}
\usage{
IPD_data
}
\description{
This dataset contains reconstructed individual patient data (IPD) from
published KM survival curve. The dataset includes information on treatment
type, treatment code, survival curve type, event times, and event occurrences
for each patient.
}
\keyword{datasets}
