% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parametric_models}
\alias{parametric_models}
\title{Fitted Survival Models}
\format{
An object of class \code{list} of length 4.
}
\usage{
parametric_models
}
\description{
This list contains the results of fitting Gompertz survival models to four
different curves: \code{GD2.EFS}, \code{TT.EFS}, \code{GD2.OS}, and
\code{TT.OS}. Each element in the list contains detailed information about
the fitted model, including the call, parameters, data used, and results.
}
\keyword{datasets}
