% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cumulative_survival.R
\name{predict_cumulative_survival}
\alias{predict_cumulative_survival}
\title{Predict Cumulative Survival}
\usage{
predict_cumulative_survival(models_fit, params)
}
\arguments{
\item{models_fit}{A named list of fitted survival models.}

\item{params}{A list of model parameters including the list of parameters in
the list \code{\link{l_psm_parameters}} in addition to:
\itemize{
  \item \code{time_horizon}: The time horizon for the model in years.
  \item \code{cycle_length}: The length of a model cycle measured in years.
  \item \code{disc_rate_costs}: The annual discount rate for incurred costs.
  \item \code{disc_rate_qalys}: The annual discount rate for accrued QALYs.
}}
}
\value{
A data frame containing the predicted cumulative survival curves with
columns for time, treatment, and survival probabilities for different
endpoints.
}
\description{
This function predicts cumulative survival curves using fitted survival
models over a specified time horizon.
}
\examples{
\dontrun{
# Load the fitted Gompertz model parameters
models_fit <- NeuroblastomaPSM::parametric_models

# Define model parameters
params <- c(
  time_horizon = 10,
  cycle_length = 1/12,
  disc_rate_costs = 0.035,
  disc_rate_qalys = 0.015,
  NeuroblastomaPSM::l_psm_parameters
)

# Predict cumulative survival
df_survival_curves_long <- NeuroblastomaPSM::predict_cumulative_survival(
  models_fit = models_fit,
  params = params
)

rbind(
  head(df_survival_curves_long, n = 5),
  tail(df_survival_curves_long, n = 5)
)
}
}
