% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_survival_curve.R
\name{predict_survival_curve}
\alias{predict_survival_curve}
\title{Predict Cumulative Survival Curve}
\usage{
predict_survival_curve(surv_mod, treatment_name, end_point, time_points)
}
\arguments{
\item{surv_mod}{A survival model object from which to extract survival
probabilities.}

\item{treatment_name}{Character string specifying the treatment name
associated with the survival model.}

\item{end_point}{Character string indicating the endpoint the survival model
is based on.}

\item{time_points}{Numeric vector of time points at which survival
probabilities should be calculated.}
}
\value{
A data frame with survival probabilities at specified time points,
        labeled with treatment and end point information.
}
\description{
Predict survival probabilities at specified time points from a survival model
object, and labels them with the treatment and endpoint information.
}
\examples{
\dontrun{
fit <- survival::survfit(
    survival::Surv(eventtime, event) ~ 1,
    data = IPD_data[IPD_data$trt_cd == "TT",]
)

survival_curve <- predict_survival_curve(
    surv_mod = fit,
    treatment_name = "Isotretinoin",
    end_point = "OS",
    time_points = seq(0, 10, 1/12)
)
}
}
