% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_psa.R
\name{sample_psa_data}
\alias{sample_psa_data}
\title{Generate Probabilistic Sensitivity Analysis (PSA) Data}
\usage{
sample_psa_data(v_psa_params, l_psa_params, n_sim)
}
\arguments{
\item{v_psa_params}{A character vector of PSA parameter names.}

\item{l_psa_params}{A list containing all model PSA parameters. The list is
expected to be either \code{\link{l_psa_parameters}} or of identical
structure.}

\item{n_sim}{An integer specifying the number of simulations to run.}
}
\value{
A data frame where each column represents a parameter, with values
sampled according to the specified distributions and arguments.
}
\description{
This function generates PSA data based on provided parameter names,
distributions, and distribution arguments.
}
\examples{
\dontrun{
# Generate PSA data
df_psa <- sample_psa_data(
 v_psa_params = names(NeuroblastomaPSM::l_psa_parameters),
 l_psa_params = NeuroblastomaPSM::l_psa_parameters,
 n_sim = 100
)

head(df_psa)

df_psa2 <- sample_psa_data(
 v_psa_params = c("u_EFS", "cost_laboratory"),
 l_psa_params = NeuroblastomaPSM::l_psa_parameters,
 n_sim = 100
)

head(df_psa2)
}
}
