% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_hazard_rate.R
\name{scale_hazard_by_timeUnit}
\alias{scale_hazard_by_timeUnit}
\title{Scale Hazard Rate by Time Unit}
\usage{
scale_hazard_by_timeUnit(v_hazard_rate, relative_time_unit)
}
\arguments{
\item{relative_time_unit}{The relative time at which the hazard rate is
predicted.}

\item{hazard_rate}{The predicted hazard rate.}
}
\value{
The probability of an event occurring within the specified time
period.
}
\description{
This function calculates the probability of an event occurring within a
specified time period using the predicted hazard rate and the relative time
unit at which the hazard rate is predicted.
}
\details{
The relationship between the hazard rate (\eqn{h}) and the probability
(\eqn{P}) of an event occurring over a time period (\eqn{t}) is given by:

\deqn{P(t) = 1 - (1 - h)^t}

where:
\itemize{
  \item \eqn{h} is the constant hazard rate.
  \item \eqn{t} is the time period.
}

The formula derives from the discrete-time approximation of survival analysis:
\itemize{
  \item The probability of surviving through a single time unit given a
  constant hazard rate \eqn{h} is \eqn{1 - h}.
  \item The probability of surviving through \eqn{t} time units is
  \eqn{(1 - h)^t}.
  \item The probability of the event occurring within \eqn{t} time units is
  the complement of the survival probability: \eqn{P(t) = 1 - (1 - h)^t}.
}
}
\examples{
\dontrun{
scaled_probability <- NeuroblastomaPSM::scale_hazard_by_timeUnit(
  v_hazard_rate = c(0.05, 0.055),
  relative_time_unit = 7/365
)

scaled_probability
}
}
