% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_economic_analysis.R
\name{calculate_treatment_costs}
\alias{calculate_treatment_costs}
\title{Calculate Treatment Costs}
\usage{
calculate_treatment_costs(params)
}
\arguments{
\item{params}{A list of model parameters including:
\itemize{
  \item \code{time_horizon}: The time horizon for the model in years.
  \item \code{cycles_per_year}: The number of cycles per year.
  \item \code{body_weight}: The body weight of the patient in kilograms.
  \item \code{unit_GD2_mg}: The dosage unit for GD2 in milligrams.
  \item \code{unit_GD2_price}: The price per unit dosage of GD2.
  \item \code{unit_TT_mg}: The dosage unit for TT in milligrams.
  \item \code{unit_TT_price}: The price per unit dosage of TT.
  \item \code{c_PPS}: The costs associated with post-progression survival.
  \item \code{u_EFS}: The utility associated with event-free survival.
  \item \code{u_PPS}: The utility associated with post-progression survival.
  \item \code{disc_rate_costs}: The annual discount rate for incurred costs.
  \item \code{disc_rate_qalys}: The annual discount rate for accrued QALYs.
}}
}
\value{
A matrix of treatment costs for the first year, with columns for GD2
and TT costs over the specified time points.
}
\description{
This function calculates the treatment costs over a given time horizon, based
on the provided model parameters.
}
\examples{
\dontrun{
# Define model parameters
params <- list(
  time_horizon = 10,
  cycles_per_year = 12,
  body_weight = 15,
  unit_GD2_mg = 20,
  unit_GD2_price = 2000,
  unit_TT_mg = 10,
  unit_TT_price = 3000,
  c_PPS = 4000
)

# Calculate treatment costs
l_treatment_costs <- calculate_treatment_costs(params = params)
View(l_treatment_costs)
}
}
