% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{l_dsa_parameters}
\alias{l_dsa_parameters}
\title{Deterministic Sensitivity Analysis (DSA) parameters}
\format{
A list with the following components:
\describe{
  \item{body_weight}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{GD2_unit_price}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{TT_unit_price}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{prob_fever}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{prob_diarrhea}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{prob_vomiting}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{prob_infection}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{prob_hypersensitivity_reaction}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{prob_capillary_leak}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{cost_fever}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{cost_diarrhea}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{cost_vomiting}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{cost_infection}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{cost_hypersensitivity_reaction}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{cost_capillary_leak}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{cost_pain}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{cost_laboratory}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{cost_hospitalisation}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{Temo_unit_price}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{Iri_unit_price}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{u_EFS}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
  \item{u_PPS}{A list with the following elements:
    \describe{
      \item{lower_ci}{Lower bound of the 95\% confidence interval.}
      \item{upper_ci}{Upper bound of the 95\% confidence interval.}
    }
  }
}
}
\usage{
l_dsa_parameters
}
\description{
This list contains the parameters used for the deterministic sensitivity
analysis (DSA).
Each parameter includes its lower and upper bounds of the 95\% confidence
interval.
}
\keyword{datasets}
