% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_psm.R
\name{run_psm}
\alias{run_psm}
\title{Run Partitioned Survival Model (PSM)}
\usage{
run_psm(models_fit = NeuroblastomaPSM::parametric_models, l_params)
}
\arguments{
\item{models_fit}{A named list of fitted survival models.}

\item{l_params}{A list of model parameters including the list of parameters
in the list \code{\link{l_psm_parameters}} in addition to:
\itemize{
  \item \code{time_horizon}: The time horizon for the model in years.
  \item \code{cycle_length}: The length of a model cycle measured in years.
  \item \code{disc_rate_costs}: The annual discount rate for incurred costs.
  \item \code{disc_rate_qalys}: The annual discount rate for accrued QALYs.
}}
}
\value{
A vector containing two scalars: discounted costs and discounted
QALYs for each treatment.
}
\description{
This function runs the Partitioned Survival Model (PSM) that estimates the
cost-effectiveness of 'Isotretinoin' vs 'Dinutuximab β' in the treatment of
Neuroblastoma in children.
}
\examples{
\dontrun{
# Load the fitted Gompertz model parameters
models_fit <- NeuroblastomaPSM::parametric_models

# Define model parameters
params <- c(
  time_horizon = 15,
  cycle_length = 1/12,
  disc_rate_costs = 0.035,
  disc_rate_qalys = 0.015,
  NeuroblastomaPSM::l_psm_parameters
)

# Run the Partitioned Survival Model
v_psm_results <- NeuroblastomaPSM::run_psm(
  models_fit = NeuroblastomaPSM::parametric_models,
  l_params = params
)

v_psm_results

# Run the Partitioned Survival Model
v_psm_results <- NeuroblastomaPSM::run_psm(
  l_params = params
)

v_psm_results
}
}
